Исправлено замечание: Модальные окна имеют внутри порталы. Их можно писать где угодно. Удобно сделать это рядом с кнопкой вызова модального окна. Так не придется прокидывать лишние функции в пропсы.
