+++ src/components/app/app.tsx
@@ -4,9 +4,6 @@ import AppHeader from "../app-header/app-header";
 import BurgerIngredients from "../burger-ingredients/burger-ingredients";
 import BurgerConstructor from "../burger-constructor/burger-constructor";
 import { BurgerIngredient } from "../../types/types";
-import Modal from "../modal/modal";
-import IngredientDetails from "../modal/ingredient-details/ingredient-details";
-import OrderDetails from "../modal/order-details/order-details";
 
 const API_URL = "https://norma.nomoreparties.space/api/ingredients";
 
@@ -14,27 +11,6 @@ function App() {
   const [ingredients, setIngredients] = useState<BurgerIngredient[]>([]);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(null);
-  const [isIngredientModalOpen, setIngredientModalOpen] = useState(false);
-  const [isOrderModalOpen, setOrderModalOpen] = useState(false);
-  const [selectedIngredient, setSelectedIngredient] =
-    useState<BurgerIngredient | null>(null);
-
-  const openIngredientModal = (ingredient: BurgerIngredient) => {
-    setSelectedIngredient(ingredient);
-    setIngredientModalOpen(true);
-  };
-
-  const closeIngredientModal = () => {
-    setIngredientModalOpen(false);
-  };
-
-  const openOrderModal = () => {
-    setOrderModalOpen(true);
-  };
-
-  const closeOrderModal = () => {
-    setOrderModalOpen(false);
-  };
 
   const getIngredients = async () => {
     try {
@@ -75,29 +51,11 @@ function App() {
 
           <div className={styles.content}>
             <main className={styles.main}>
-              <BurgerIngredients
-                data={ingredients}
-                onIngredientClick={openIngredientModal}
-              />
-              <BurgerConstructor
-                data={ingredients}
-                onOrderClick={openOrderModal}
-              />
+              <BurgerIngredients data={ingredients} />
+              <BurgerConstructor data={ingredients} />
             </main>
           </div>
         </>
-      )}
-
-      {isIngredientModalOpen && selectedIngredient && (
-        <Modal title="Детали ингредиента" onClose={closeIngredientModal}>
-          <IngredientDetails ingredient={selectedIngredient} />
-        </Modal>
-      )}
-
-      {isOrderModalOpen && (
-        <Modal title="" onClose={closeOrderModal}>
-          <OrderDetails />
-        </Modal>
       )}
     </>
   );
