+++ src/components/burger-constructor/burger-constructor.tsx
@@ -1,13 +1,22 @@
-import React from "react";
+import React, { useState } from "react";
 import styles from "./burger-constructor.module.css";
 import BurgerConstructorList from "./burger-constructor-list/burger-constructor-list";
 import TotalPrice from "../total-price/total-price";
 import { Button } from "@ya.praktikum/react-developer-burger-ui-components";
 import { BurgerIngredientsProps } from "../../types/types";
+import Modal from "../modal/modal";
+import OrderDetails from "../modal/order-details/order-details";
 
-const BurgerConstructor: React.FC<
-  BurgerIngredientsProps & { onOrderClick: () => void }
-> = ({ data, onOrderClick }) => {
+const BurgerConstructor: React.FC<BurgerIngredientsProps> = ({ data }) => {
+  const [isOrderModalOpen, setOrderModalOpen] = useState(false);
+
+  const openOrderModal = () => {
+    setOrderModalOpen(true);
+  };
+
+  const closeOrderModal = () => {
+    setOrderModalOpen(false);
+  };
   return (
     <section className={styles.column}>
       <BurgerConstructorList data={data} />
@@ -17,10 +26,15 @@ const BurgerConstructor: React.FC<
           htmlType="button"
           type="primary"
           size="medium"
-          onClick={onOrderClick}
+          onClick={openOrderModal}
         >
           Оформить заказ
         </Button>
+        {isOrderModalOpen && (
+          <Modal title="" onClose={closeOrderModal}>
+            <OrderDetails />
+          </Modal>
+        )}
       </div>
     </section>
   );
