+++ src/components/burger-ingredients/burger-ingredients-list/burger-ingredients-list.tsx
@@ -1,26 +1,18 @@
 import React from "react";
 import styles from "./burger-ingredients-list.module.css";
 import BurgerIngredientsItem from "./burger-ingredients-item/burger-ingredients-item";
-import {
-  BurgerIngredient,
-  BurgerIngredientsListProps,
-} from "../../../types/types";
+import { BurgerIngredientsListProps } from "../../../types/types";
 
-const BurgerIngredientsList: React.FC<
-  BurgerIngredientsListProps & {
-    onIngredientClick: (ingredient: BurgerIngredient) => void;
-  }
-> = ({ data, type, onIngredientClick }) => {
+const BurgerIngredientsList: React.FC<BurgerIngredientsListProps> = ({
+  data,
+  type,
+}) => {
   const items = data.filter((item) => item.type === type);
 
   return (
     <ul className={styles.list}>
       {items.map((item) => (
-        <BurgerIngredientsItem
-          key={item._id}
-          {...item}
-          onIngredientClick={onIngredientClick}
-        />
+        <BurgerIngredientsItem key={item._id} {...item} />
       ))}
     </ul>
   );
